requirements.txt
bcrypt==3.2.0
cffi==1.14.2
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
Flask-PyMongo==2.3.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pycparser==2.20
pymongo[tls,srv]==3.11.0
six==1.15.0
Werkzeug==1.0.1
requests==2.24.0
pytest==6.0.2
python-crontab